requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
yfinance>=0.2.28
newsapi-python>=0.2.7
praw>=7.7.0
transformers>=4.30.0
torch>=2.0.0
ta>=0.11.0
pandas-ta>=0.3.14b
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
shap>=0.42.0
python-dotenv>=1.0.0
PyYAML>=6.0
tqdm>=4.65.0
requests>=2.31.0
streamlit>=1.28.0
pytest>=7.4.0